requirements.txt
flask==3.0.3
gunicorn==22.0.0
tensorflow==2.15.1
numpy==1.26.4
scikit-learn==1.5.0
google-cloud-storage==2.17.0
python-dotenv==1.0.1
